requirements.txt
# Gemini Subtitle Matcher - Python Dependencies
# Minimal dependencies for lightweight desktop application

# Official Google Gemini API SDK
google-genai>=0.1.0

# Note: tkinter is included with Python standard library
# No additional GUI dependencies required
